models of # ticks infected on mice and power calculations
MODELS: outcome of # infected ticks on each mouse
A poisson would be a reasonable rough model
A zero-inflated poisson deals with the extra zero's issue.
I ran both, results attached in the word doc (and pasted below),
issues: neither takes into account the difference in total number of
ticks examined, will have to think about that.

POWER: I did power calcs by permutation, resampling from the data.
The .pdf shows the power for sample sizes of 30 to 100 per arm.
The black line is the poisson regression. With 60 mice per arm Power is
~ 85%
The power is a bit lower for the zero-inflated model (blue line) but
not too bad with 60 animals per arm.

CODE:
 in the attached .R doc.
 I had to manipulate the data to bring into R, .csv for the code
attached
you'll have to do the usual switches of working directory etc.
I've started keeping all codes in Github- might be a good idea as
things accumulate.